+++ packages/std/src/exports.rs
@@ -15,7 +15,7 @@ use crate::imports::{ExternalApi, ExternalStorage};
 use crate::memory::{alloc, consume_region, release_buffer};
 use crate::serde::{from_slice, to_vec};
 use crate::traits::Extern;
-use crate::{ContractResult, Env, QueryResult, Response};
+use crate::{Env, HandleResponse, HandleResult, InitResponse, InitResult, QueryResult};
 
 /// cosmwasm_vm_version_* exports mark which Wasm VM interface level this contract is compiled for.
 /// They can be checked by cosmwasm_vm.
@@ -41,13 +41,17 @@ pub extern "C" fn deallocate(pointer: *mut c_void) {
 
 /// do_init should be wrapped in an external "C" export, containing a contract-specific function as arg
 pub fn do_init<T: DeserializeOwned + JsonSchema>(
-    init_fn: &dyn Fn(&mut Extern<ExternalStorage, ExternalApi>, Env, T) -> Result<Response, Error>,
+    init_fn: &dyn Fn(
+        &mut Extern<ExternalStorage, ExternalApi>,
+        Env,
+        T,
+    ) -> Result<InitResponse, Error>,
     env_ptr: *mut c_void,
     msg_ptr: *mut c_void,
 ) -> *mut c_void {
     match _do_init(init_fn, env_ptr, msg_ptr) {
         Ok(res) => res,
-        Err(err) => make_error_c_string(err),
+        Err(err) => make_init_error_c_string(err),
     }
 }
 
@@ -57,13 +61,13 @@ pub fn do_handle<T: DeserializeOwned + JsonSchema>(
         &mut Extern<ExternalStorage, ExternalApi>,
         Env,
         T,
-    ) -> Result<Response, Error>,
+    ) -> Result<HandleResponse, Error>,
     env_ptr: *mut c_void,
     msg_ptr: *mut c_void,
 ) -> *mut c_void {
     match _do_handle(handle_fn, env_ptr, msg_ptr) {
         Ok(res) => res,
-        Err(err) => make_error_c_string(err),
+        Err(err) => make_handle_error_c_string(err),
     }
 }
 
@@ -79,7 +83,11 @@ pub fn do_query<T: DeserializeOwned + JsonSchema>(
 }
 
 fn _do_init<T: DeserializeOwned + JsonSchema>(
-    init_fn: &dyn Fn(&mut Extern<ExternalStorage, ExternalApi>, Env, T) -> Result<Response, Error>,
+    init_fn: &dyn Fn(
+        &mut Extern<ExternalStorage, ExternalApi>,
+        Env,
+        T,
+    ) -> Result<InitResponse, Error>,
     env_ptr: *mut c_void,
     msg_ptr: *mut c_void,
 ) -> Result<*mut c_void, Error> {
@@ -89,7 +97,7 @@ fn _do_init<T: DeserializeOwned + JsonSchema>(
     let msg: T = from_slice(&msg)?;
     let mut deps = make_dependencies();
     let res = init_fn(&mut deps, env, msg)?;
-    let json = to_vec(&ContractResult::Ok(res))?;
+    let json = to_vec(&InitResult::Ok(res))?;
     Ok(release_buffer(json))
 }
 
@@ -98,7 +106,7 @@ fn _do_handle<T: DeserializeOwned + JsonSchema>(
         &mut Extern<ExternalStorage, ExternalApi>,
         Env,
         T,
-    ) -> Result<Response, Error>,
+    ) -> Result<HandleResponse, Error>,
     env_ptr: *mut c_void,
     msg_ptr: *mut c_void,
 ) -> Result<*mut c_void, Error> {
@@ -109,7 +117,7 @@ fn _do_handle<T: DeserializeOwned + JsonSchema>(
     let msg: T = from_slice(&msg)?;
     let mut deps = make_dependencies();
     let res = handle_fn(&mut deps, env, msg)?;
-    let json = to_vec(&ContractResult::Ok(res))?;
+    let json = to_vec(&HandleResult::Ok(res))?;
     Ok(release_buffer(json))
 }
 
@@ -126,8 +134,13 @@ fn _do_query<T: DeserializeOwned + JsonSchema>(
     Ok(release_buffer(json))
 }
 
-fn make_error_c_string<T: Display>(error: T) -> *mut c_void {
-    let v = to_vec(&ContractResult::Err(error.to_string())).unwrap();
+fn make_init_error_c_string<T: Display>(error: T) -> *mut c_void {
+    let v = to_vec(&InitResult::Err(error.to_string())).unwrap();
+    release_buffer(v)
+}
+
+fn make_handle_error_c_string<T: Display>(error: T) -> *mut c_void {
+    let v = to_vec(&HandleResult::Err(error.to_string())).unwrap();
     release_buffer(v)
 }